% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunr-package.R
\docType{package}
\name{lacunr.custom-package}
\alias{lacunr.custom}
\alias{lacunr.custom-package}
\title{lacunr.custom: Fast 3D Lacunarity for Voxel Data}
\description{
Calculates 3D lacunarity from voxel data. It is designed for use with point clouds generated from Light Detection And Ranging (LiDAR) scans in order to measure the spatial heterogeneity of 3-dimensional structures such as forest stands. It provides fast 'C++' functions to efficiently bin point cloud data into voxels and calculate lacunarity using different variants of the gliding-box algorithm originated by Allain & Cloitre (1991) \doi{10.1103/PhysRevA.44.3552}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElliottSmeds/lacunr}
  \item Report bugs at \url{https://github.com/ElliottSmeds/lacunr/issues}
}

}
\author{
\strong{Maintainer}: Elliott Smeds \email{elliott.alfred93@gmail.com} (\href{https://orcid.org/0000-0003-1054-7491}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item J. Antonio Guzmán Q. (Author of original version of voxelize() function) [copyright holder]
}

}
\keyword{internal}
